Moved Copy text functionality into own component, old use from components
