+++ website/src/components/copy/styles.less
@@ -1,0 +1,9 @@
+.copy {
+  &__popover {
+    padding: 0.5rem;
+  }
+
+  &__copyknapp {
+    margin-left: 0.5em;
+  }
+}
