+++ website/src/components/layout/templates/styles.less
@@ -81,16 +81,6 @@
   }
 }
 
-.technical {
-  &__popover {
-    padding: 0.7rem;
-  }
-
-  &__copyknapp {
-    margin-left: 0.2em;
-  }
-}
-
 #beta-component-modal {
   padding: 3rem 3.5rem 2.5rem;
   max-width: 500px;
