+++ website/src/components/proptable/ModuleBrowser.tsx
@@ -1,13 +1,11 @@
 import classnames from "classnames";
-import { Flatknapp } from "nav-frontend-knapper";
 import Panel from "nav-frontend-paneler";
-import Popover, { PopoverOrientering } from "nav-frontend-popover";
 import { Select } from "nav-frontend-skjema";
 import { Systemtittel, Undertittel } from "nav-frontend-typografi";
 import React, { useEffect, useState } from "react";
 import { useProps } from "../../useProps";
-import { CopyIcon } from "../assets/images/svg";
 import Code from "../code/Code";
+import Copy from "../copy/Copy";
 import PropTable from "./PropTable";
 import "./styles.less";
 
@@ -16,7 +14,6 @@ const ModuleBrowser = ({ context, ...props }) => {
   const modules = useProps(context.source);
 
   const [activeModule, setActiveModule] = useState<number>(0);
-  const [anchor, setAnchor] = useState(undefined);
 
   useEffect(() => {
     setActiveModule(getInitialActiveModule());
@@ -41,16 +38,6 @@ const ModuleBrowser = ({ context, ...props }) => {
         ? modules[activeModule].name
         : `{ ${modules[activeModule].name} }`;
     return `import ${format} from '${context.name}';`;
-  };
-
-  const copyContent = (e, content) => {
-    setAnchor(anchor ? undefined : e.currentTarget);
-    const textArea = document.createElement("textarea");
-    textArea.value = content;
-    document.body.appendChild(textArea);
-    textArea.select();
-    document.execCommand("Copy");
-    textArea.remove();
   };
 
   return (
@@ -103,24 +90,10 @@ const ModuleBrowser = ({ context, ...props }) => {
 
             <div className="module-browser--innline">
               <Undertittel className="first">Import</Undertittel>
-              <Flatknapp
-                className="module-browser__copyknapp"
-                aria-label="Kopier import til utklippstavle"
-                // eslint-disable-next-line max-len
-                onClick={(e) => copyContent(e, generateImportStatement())}
-                kompakt
-              >
-                <CopyIcon />
-              </Flatknapp>
+              <Copy copyText={generateImportStatement()} />
             </div>
             <Code className="language-jsx">{generateImportStatement()}</Code>
-            <Popover
-              orientering={PopoverOrientering.Over}
-              ankerEl={anchor}
-              onRequestClose={() => setAnchor(undefined)}
-            >
-              <p className="module-browser__popover"> Kopiert! </p>
-            </Popover>
+
             <Undertittel>React props</Undertittel>
 
             <PropTable moduleProps={modules[activeModule]} />
