+++ website/src/components/proptable/styles.less
@@ -84,14 +84,6 @@
     }
   }
 
-  &__popover {
-    padding: 0.7rem;
-  }
-
-  &__copyknapp {
-    margin-left: 0.2em;
-  }
-
   &--innline {
     display: flex;
     flex-direction: row;
